test: remove outdated unit tests and update mocks for Home, Project, Blog, and About pages
